# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "검색",
      "description": "줄거리 의미 검색 (dense + sparse 하이브리드) · 유사 영화"
    },
    {
      "name": "영화",
      "description": "상세 · 탐색 · 장르 · 인물 조회 (PG 기준)"
    }
  ],
  "paths": {
    "/cineseek/search": {
      "get": {
        "tags": [
          "검색"
        ],
        "summary": "자연어로 영화 검색",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "검색어 (줄거리 묘사)",
            "required": true,
            "schema": {
              "type": "string",
              "description": "검색어 (줄거리 묘사)",
              "example": "감옥에서 탈출하는 이야기",
              "minLength": 1
            },
            "example": "감옥에서 탈출하는 이야기"
          },
          {
            "name": "genreId",
            "in": "query",
            "description": "장르 id (반복 파라미터, OR)",
            "required": false,
            "schema": {
              "type": "array",
              "description": "장르 id (반복 파라미터, OR)",
              "example": [
                80
              ],
              "items": {
                "type": "integer",
                "format": "int64",
                "description": "장르 id (반복 파라미터, OR)",
                "example": [
                  80
                ]
              }
            },
            "example": [
              80
            ]
          },
          {
            "name": "yearMin",
            "in": "query",
            "description": "개봉 연도 하한",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "개봉 연도 하한",
              "example": 2000
            },
            "example": 2000
          },
          {
            "name": "yearMax",
            "in": "query",
            "description": "개봉 연도 상한",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "개봉 연도 상한",
              "example": 2020
            },
            "example": 2020
          },
          {
            "name": "ratingMin",
            "in": "query",
            "description": "평점 하한 (0~10)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "description": "평점 하한 (0~10)",
              "example": 7.5
            },
            "example": 7.5
          },
          {
            "name": "runtimeMin",
            "in": "query",
            "description": "러닝타임 하한 (분)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "러닝타임 하한 (분)",
              "example": 90,
              "maximum": 1000,
              "minimum": 0
            },
            "example": 90
          },
          {
            "name": "runtimeMax",
            "in": "query",
            "description": "러닝타임 상한 (분)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "러닝타임 상한 (분)",
              "example": 150,
              "maximum": 1000,
              "minimum": 0
            },
            "example": 150
          },
          {
            "name": "voteCountMin",
            "in": "query",
            "description": "투표 수 하한 — 숨은 명작 탐색용",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "투표 수 하한 — 숨은 명작 탐색용",
              "example": 50,
              "maximum": 1000000,
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "voteCountMax",
            "in": "query",
            "description": "투표 수 상한 — 숨은 명작 탐색용",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "투표 수 상한 — 숨은 명작 탐색용",
              "example": 500,
              "maximum": 1000000,
              "minimum": 0
            },
            "example": 500
          },
          {
            "name": "directorId",
            "in": "query",
            "description": "감독 person id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "감독 person id",
              "example": 525
            },
            "example": 525
          },
          {
            "name": "castId",
            "in": "query",
            "description": "배우 person id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "배우 person id",
              "example": 3895
            },
            "example": 3895
          },
          {
            "name": "limit",
            "in": "query",
            "description": "결과 수 (1~50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "결과 수 (1~50)",
              "example": 20,
              "maximum": 50,
              "minimum": 1
            },
            "example": 20
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 결과 수",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "건너뛸 결과 수",
              "example": 0,
              "minimum": 0
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/people": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "감독·배우 자동완성",
        "operationId": "people",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "이름 접두어",
            "required": true,
            "schema": {
              "type": "string",
              "description": "이름 접두어",
              "example": "봉",
              "minLength": 1
            },
            "example": "봉"
          },
          {
            "name": "role",
            "in": "query",
            "description": "역할 (director | cast)",
            "required": false,
            "schema": {
              "type": "string",
              "description": "역할 (director | cast)",
              "example": "director",
              "pattern": "director|cast"
            },
            "example": "director"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/people/{id}": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "인물 이름 조회 (필터 칩 라벨 복원용)",
        "operationId": "person",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "역할 (director | cast)",
            "required": false,
            "schema": {
              "type": "string",
              "description": "역할 (director | cast)",
              "example": "director",
              "pattern": "director|cast"
            },
            "example": "director"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/people/{id}/filmography": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "배우 필모그래피 — 저장된 출연작을 최신순·평점순으로",
        "operationId": "filmography",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "정렬 (release | rating) — 기본 최신 개봉순",
            "required": false,
            "schema": {
              "type": "string",
              "description": "정렬 (release | rating) — 기본 최신 개봉순",
              "example": "release",
              "pattern": "release|rating"
            },
            "example": "release"
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 (0부터)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "페이지 (0부터)",
              "example": 0,
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지 크기 (1~50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "페이지 크기 (1~50)",
              "example": 20,
              "maximum": 50,
              "minimum": 1
            },
            "example": 20
          },
          {
            "name": "excludeMovieId",
            "in": "query",
            "description": "제외할 영화 id — 영화 상세의 '주요 배우 다른 작품'에서 현재 영화 제외",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "제외할 영화 id — 영화 상세의 '주요 배우 다른 작품'에서 현재 영화 제외",
              "example": 155
            },
            "example": 155
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FilmographyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/movies": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "장르·인물 조건으로 영화 탐색",
        "operationId": "explore",
        "parameters": [
          {
            "name": "genreId",
            "in": "query",
            "description": "장르 id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "장르 id",
              "example": 28
            },
            "example": 28
          },
          {
            "name": "sort",
            "in": "query",
            "description": "정렬 (rating | release | vote_count)",
            "required": false,
            "schema": {
              "type": "string",
              "description": "정렬 (rating | release | vote_count)",
              "example": "rating",
              "pattern": "rating|release|vote_count"
            },
            "example": "rating"
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 (0부터)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "페이지 (0부터)",
              "example": 0,
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지 크기 (1~50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "페이지 크기 (1~50)",
              "example": 20,
              "maximum": 50,
              "minimum": 1
            },
            "example": 20
          },
          {
            "name": "directorId",
            "in": "query",
            "description": "감독 person id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "감독 person id",
              "example": 525
            },
            "example": 525
          },
          {
            "name": "castId",
            "in": "query",
            "description": "배우 person id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "배우 person id",
              "example": 3895
            },
            "example": 3895
          },
          {
            "name": "ratingMin",
            "in": "query",
            "description": "평점 하한 (0~10)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "description": "평점 하한 (0~10)",
              "example": 7.5,
              "maximum": 10.0,
              "minimum": 0.0
            },
            "example": 7.5
          },
          {
            "name": "voteCountMin",
            "in": "query",
            "description": "최소 투표 수 — 명시하면 평점순 암시 하한(1000) 대신 적용",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "최소 투표 수 — 명시하면 평점순 암시 하한(1000) 대신 적용",
              "example": 50,
              "maximum": 1000000,
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "voteCountMax",
            "in": "query",
            "description": "최대 투표 수 (숨은 명작 탐색용)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "최대 투표 수 (숨은 명작 탐색용)",
              "example": 500,
              "maximum": 1000000,
              "minimum": 0
            },
            "example": 500
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MovieListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/movies/{id}": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "영화 상세",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/movies/{id}/similar": {
      "get": {
        "tags": [
          "검색"
        ],
        "summary": "줄거리가 비슷한 영화",
        "operationId": "similar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "결과 수 (1~20)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "결과 수 (1~20)",
              "example": 8,
              "maximum": 20,
              "minimum": 1
            },
            "example": 8
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cineseek/genres": {
      "get": {
        "tags": [
          "영화"
        ],
        "summary": "장르 목록",
        "operationId": "genres",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenresResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GenreResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "TMDB 장르 id",
            "example": 28
          },
          "name": {
            "type": "string"
          },
          "nameKo": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PageResponse": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "미제공 — 하이브리드 검색 전체 count 비용"
          }
        }
      },
      "SearchItemResponse": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "releaseYear": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "voteCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "posterPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          },
          "score": {
            "type": "number",
            "format": "double",
            "description": "검색은 RRF 점수, 유사 영화는 코사인 유사도",
            "example": 0.032
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "검색 결과 (점수 내림차순)",
            "items": {
              "$ref": "#/components/schemas/SearchItemResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          }
        }
      },
      "PeopleResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonItemResponse"
            }
          }
        }
      },
      "PersonItemResponse": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "knownFor": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MovieCardResponse": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "releaseYear": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "TMDB 평점 (0~10)",
            "example": 8.7
          },
          "voteCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "posterPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          }
        }
      },
      "MovieListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCardResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          }
        }
      },
      "CastMemberResponse": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "character": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "MovieDetailResponse": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "releaseYear": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "runtime": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "voteCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "posterPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          },
          "directors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            }
          },
          "cast": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CastMemberResponse"
            }
          }
        }
      },
      "PersonResponse": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "GenresResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          }
        }
      },
      "FilmographyPersonResponse": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "integer",
            "format": "int64",
            "description": "TMDB person id",
            "example": 3895
          },
          "name": {
            "type": "string",
            "description": "이름",
            "example": "크리스찬 베일"
          },
          "profilePath": {
            "type": [
              "string",
              "null"
            ],
            "description": "TMDB 프로필 이미지 경로",
            "example": "/fuTEtOeIwXiqtVpIoL0GyyWcnxh.jpg"
          },
          "filmoState": {
            "type": "string",
            "description": "필모그래피 수집 상태 (NONE | COLLECTING | COMPLETE)",
            "example": "COMPLETE",
            "enum": [
              "NONE",
              "COLLECTING",
              "COMPLETE"
            ]
          },
          "filmoCheckedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "필모그래피 마지막 갱신 확인 시각",
            "example": "2026-10-01T00:00:00Z"
          },
          "totalWorks": {
            "type": "integer",
            "format": "int64",
            "description": "총 출연 작품 수 (movie_cast 행 수)",
            "example": 47
          }
        }
      },
      "FilmographyItemResponse": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64",
            "description": "영화 id",
            "example": 155
          },
          "title": {
            "type": "string",
            "description": "제목",
            "example": "다크 나이트"
          },
          "originalTitle": {
            "type": [
              "string",
              "null"
            ],
            "description": "원제",
            "example": "The Dark Knight"
          },
          "releaseYear": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "개봉 연도",
            "example": 2008
          },
          "rating": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "TMDB 평점 (0~10)",
            "example": 8.5
          },
          "voteCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "TMDB 투표 수",
            "example": 30000
          },
          "posterPath": {
            "type": [
              "string",
              "null"
            ],
            "description": "포스터 이미지 경로",
            "example": "/qJ2tW6WMUDux911r6m7haRef0WH.jpg"
          },
          "genres": {
            "type": "array",
            "description": "장르 목록",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          },
          "character": {
            "type": [
              "string",
              "null"
            ],
            "description": "해당 배우의 배역명",
            "example": "브루스 웨인"
          }
        }
      },
      "FilmographyResponse": {
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/FilmographyPersonResponse"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilmographyItemResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          },
          "collecting": {
            "type": "boolean",
            "description": "필모그래피 백그라운드 갱신이 대기/실행 중인지 — true면 잠시 후 다시 조회해 갱신된 목록을 받는다",
            "example": false
          }
        }
      }
    }
  }
}